requirements.txt
google-generativeai
python-dotenv
requests
beautifulsoup4
feedparser
faiss-cpu
sentence-transformers
flask